Creando factura

+++ src/main/java/Facturas.java
@@ -6,9 +6,9 @@ public class Facturas {
     private Tarifa tarifa;
     private Calendar fecha_emision;
     private Calendar fecha_facturacion;
-    private int importe;
+    private double importe;
 
-    public Facturas(int codigo, Tarifa tarifa, Calendar fecha_emision, Calendar fecha_facturacion, int importe) {
+    public Facturas(int codigo, Tarifa tarifa, Calendar fecha_emision, Calendar fecha_facturacion, double importe) {
         this.codigo = codigo;
         this.tarifa = tarifa;
         this.fecha_emision = fecha_emision;
@@ -32,7 +32,7 @@ public class Facturas {
         return fecha_facturacion;
     }
 
-    public int getImporte() {
+    public double getImporte() {
         return importe;
     }
 
@@ -52,7 +52,7 @@ public class Facturas {
         this.fecha_facturacion = fecha_facturacion;
     }
 
-    public void setImporte(int importe) {
+    public void setImporte(double importe) {
         this.importe = importe;
     }
